nexum-task/requirements.txt
fastapi==0.101.1
uvicorn==0.23.2
python-jose==3.5.0
pydantic==2.1.1
mysql-connector-python==9.6.0
python-multipart==0.0.6
python-dotenv==1.0.0